+++ ethexe/common/src/crypto/mod.rs
@@ -7,4 +7,3 @@ pub use address::*;
 pub use digest::*;
 pub use keys::*;
 pub use signature::*;
-
