+++ ethexe/processor/src/handling/mod.rs
@@ -18,7 +18,8 @@
 
 use crate::Processor;
 use anyhow::{anyhow, Result};
-use ethexe_db::{BlockMetaStorage, CodesStorage, Database, OnChainStorage};
+use ethexe_common::db::{BlockMetaStorage, CodesStorage, OnChainStorage};
+use ethexe_db::Database;
 use ethexe_runtime_common::{
     state::ProgramState, InBlockTransitions, ScheduleHandler, TransitionController,
 };
