+++ ethexe/processor/src/lib.rs
@@ -20,11 +20,12 @@
 
 use anyhow::{anyhow, ensure, Result};
 use ethexe_common::{
-    db::Schedule,
+    db::CodesStorage,
     events::{BlockRequestEvent, MirrorRequestEvent},
     gear::StateTransition,
+    Schedule,
 };
-use ethexe_db::{CodesStorage, Database};
+use ethexe_db::Database;
 use ethexe_runtime_common::state::Storage;
 use gear_core::{ids::prelude::CodeIdExt, message::ReplyInfo};
 use gprimitives::{ActorId, CodeId, MessageId, H256};
