+++ ethexe/processor/src/tests.rs
@@ -17,8 +17,11 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::*;
-use ethexe_common::events::{BlockRequestEvent, MirrorRequestEvent, RouterRequestEvent};
-use ethexe_db::{BlockHeader, BlockMetaStorage, CodesStorage, MemDb, OnChainStorage};
+use ethexe_common::{
+    db::{BlockMetaStorage, CodesStorage, OnChainStorage},
+    events::{BlockRequestEvent, MirrorRequestEvent, RouterRequestEvent},
+    BlockHeader,
+};
 use ethexe_runtime_common::ScheduleRestorer;
 use gear_core::ids::{prelude::CodeIdExt, ProgramId};
 use gprimitives::{ActorId, MessageId};
@@ -65,9 +68,7 @@ fn init_new_block_from_parent(processor: &mut Processor, parent_hash: H256) -> H
 fn process_observer_event() {
     init_logger();
 
-    let db = MemDb::default();
-    let mut processor =
-        Processor::new(Database::from_one(&db)).expect("failed to create processor");
+    let mut processor = Processor::new(Database::memory()).expect("failed to create processor");
 
     let parent = init_genesis_block(&mut processor);
     let ch0 = init_new_block_from_parent(&mut processor, parent);
@@ -157,9 +158,7 @@ fn process_observer_event() {
 fn handle_new_code_valid() {
     init_logger();
 
-    let db = MemDb::default();
-    let mut processor =
-        Processor::new(Database::from_one(&db)).expect("failed to create processor");
+    let mut processor = Processor::new(Database::memory()).expect("failed to create processor");
 
     init_genesis_block(&mut processor);
 
@@ -201,9 +200,7 @@ fn handle_new_code_valid() {
 fn handle_new_code_invalid() {
     init_logger();
 
-    let db = MemDb::default();
-    let mut processor =
-        Processor::new(Database::from_one(&db)).expect("failed to create processor");
+    let mut processor = Processor::new(Database::memory()).expect("failed to create processor");
 
     init_genesis_block(&mut processor);
 
@@ -231,8 +228,7 @@ fn handle_new_code_invalid() {
 fn ping_pong() {
     init_logger();
 
-    let db = MemDb::default();
-    let mut processor = Processor::new(Database::from_one(&db)).unwrap();
+    let mut processor = Processor::new(Database::memory()).unwrap();
 
     let parent = init_genesis_block(&mut processor);
     let ch0 = init_new_block_from_parent(&mut processor, parent);
@@ -310,8 +306,7 @@ fn async_and_ping() {
     };
     let user_id = ActorId::from(10);
 
-    let db = MemDb::default();
-    let mut processor = Processor::new(Database::from_one(&db)).unwrap();
+    let mut processor = Processor::new(Database::memory()).unwrap();
 
     let parent = init_genesis_block(&mut processor);
     let ch0 = init_new_block_from_parent(&mut processor, parent);
@@ -448,8 +443,7 @@ fn many_waits() {
 
     let (_, code) = wat_to_wasm(wat);
 
-    let db = MemDb::default();
-    let mut processor = Processor::new(Database::from_one(&db)).unwrap();
+    let mut processor = Processor::new(Database::memory()).unwrap();
 
     let parent = init_genesis_block(&mut processor);
     let ch0 = init_new_block_from_parent(&mut processor, parent);
