+++ ethexe/rpc/src/apis/code.rs
@@ -17,7 +17,8 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::errors;
-use ethexe_db::{CodesStorage, Database};
+use ethexe_common::db::CodesStorage;
+use ethexe_db::Database;
 use gprimitives::H256;
 use jsonrpsee::{
     core::{async_trait, RpcResult},
